refactor: move IPatient to interface packages
